Implement lazy dictionary
