+++ sympy/functions/elementary/trigonometric.py
@@ -18,6 +18,7 @@ from sympy.functions.elementary.exponential import log, exp
 from sympy.functions.elementary.integers import floor
 from sympy.functions.elementary.miscellaneous import sqrt, Min, Max
 from sympy.functions.elementary.piecewise import Piecewise
+from sympy.functions.elementary._trigonometric_special import cos_table, migcdex
 from sympy.logic.boolalg import And
 from sympy.ntheory import factorint
 from sympy.polys.specialpolys import symmetric_poly
@@ -236,61 +237,6 @@ def _pi_coeff(arg, cycles=1):
             return cx
     elif arg.is_zero:
         return S.Zero
-
-
-@cacheit
-def _cospi257():
-    """ Express cos(pi/257) explicitly as a function of radicals
-        Based upon the equations in
-        http://math.stackexchange.com/questions/516142/how-does-cos2-pi-257-look-like-in-real-radicals
-        See also https://r-knott.surrey.ac.uk/Fibonacci/simpleTrig.html
-    """
-    def f1(a, b):
-        return (a + sqrt(a**2 + b))/2, (a - sqrt(a**2 + b))/2
-
-    def f2(a, b):
-        return (a - sqrt(a**2 + b))/2
-
-    t1, t2 = f1(-1, 256)
-    z1, z3 = f1(t1, 64)
-    z2, z4 = f1(t2, 64)
-    y1, y5 = f1(z1, 4*(5 + t1 + 2*z1))
-    y6, y2 = f1(z2, 4*(5 + t2 + 2*z2))
-    y3, y7 = f1(z3, 4*(5 + t1 + 2*z3))
-    y8, y4 = f1(z4, 4*(5 + t2 + 2*z4))
-    x1, x9 = f1(y1, -4*(t1 + y1 + y3 + 2*y6))
-    x2, x10 = f1(y2, -4*(t2 + y2 + y4 + 2*y7))
-    x3, x11 = f1(y3, -4*(t1 + y3 + y5 + 2*y8))
-    x4, x12 = f1(y4, -4*(t2 + y4 + y6 + 2*y1))
-    x5, x13 = f1(y5, -4*(t1 + y5 + y7 + 2*y2))
-    x6, x14 = f1(y6, -4*(t2 + y6 + y8 + 2*y3))
-    x15, x7 = f1(y7, -4*(t1 + y7 + y1 + 2*y4))
-    x8, x16 = f1(y8, -4*(t2 + y8 + y2 + 2*y5))
-    v1 = f2(x1, -4*(x1 + x2 + x3 + x6))
-    v2 = f2(x2, -4*(x2 + x3 + x4 + x7))
-    v3 = f2(x8, -4*(x8 + x9 + x10 + x13))
-    v4 = f2(x9, -4*(x9 + x10 + x11 + x14))
-    v5 = f2(x10, -4*(x10 + x11 + x12 + x15))
-    v6 = f2(x16, -4*(x16 + x1 + x2 + x5))
-    u1 = -f2(-v1, -4*(v2 + v3))
-    u2 = -f2(-v4, -4*(v5 + v6))
-    w1 = -2*f2(-u1, -4*u2)
-    return sqrt(sqrt(2)*sqrt(w1 + 4)/8 + S.Half)
-
-
-@cacheit
-def _cos_sqrt_cst_table_some():
-    return {
-        3: S.Half,
-        5: (sqrt(5) + 1) / 4,
-        17: sqrt((15 + sqrt(17)) / 32 + sqrt(2) * (sqrt(17 - sqrt(17)) +
-            sqrt(sqrt(2) * (-8 * sqrt(17 + sqrt(17)) - (1 - sqrt(17))
-            * sqrt(17 - sqrt(17))) + 6 * sqrt(17) + 34)) / 32),
-        257: _cospi257()
-        # 65537 is the only other known Fermat prime and the very
-        # large expression is intentionally omitted from SymPy; see
-        # https://r-knott.surrey.ac.uk/Fibonacci/simpleTrig.html
-    }
 
 
 class sin(TrigonometricFunction):
@@ -851,7 +797,6 @@ class cos(TrigonometricFunction):
 
     def _eval_rewrite_as_sqrt(self, arg, **kwargs):
         from sympy.functions.special.polynomials import chebyshevt
-        from sympy.functions.elementary.trigonometric_special import migcdex
 
         def ipartfrac(r, factors=None):
             if isinstance(r, int):
@@ -877,17 +822,16 @@ class cos(TrigonometricFunction):
         if pi_coeff is None:
             return None
 
-        if pi_coeff.is_integer:
-            # it was unevaluated
-            return self.func(pi_coeff*pi)
-
-        if not pi_coeff.is_Rational:
+        if isinstance(pi_coeff, Integer):
             return None
 
-        cst_table_some = _cos_sqrt_cst_table_some()
+        if not isinstance(pi_coeff, Rational):
+            return None
+
+        cst_table_some = cos_table()
 
         if pi_coeff.q in cst_table_some:
-            rv = chebyshevt(pi_coeff.p, cst_table_some[pi_coeff.q])
+            rv = chebyshevt(pi_coeff.p, cst_table_some[pi_coeff.q]())
             if pi_coeff.q < 257:
                 rv = rv.expand()
             return rv
@@ -912,6 +856,7 @@ class cos(TrigonometricFunction):
                     if n == 1:
                         return tuple(primes)
             return False
+
         FC = _fermatCoords(pi_coeff.q)
         if FC:
             decomp = ipartfrac(pi_coeff, FC)
